fix: MS-5 Reaper implementation - fix autoincrement and test issues
Fixed critical bugs preventing Reaper tests from passing:

1. models.py - Fix autoincrement for TenantPlan and TenantUsageDaily
   - Changed BIGINT to Integer for id fields
   - SQLite requires INTEGER PRIMARY KEY for proper autoincrement
   - PostgreSQL compatibility maintained (Integer → SERIAL)

2. conftest.py - Add multi-threading support for race condition tests
   - Added db_engine fixture for shared engine across threads
   - Modified db_session to use shared engine
   - Enables proper session isolation in concurrent tests

3. test_reaper.py - Fix race condition test issues
   - Fixed BudgetScripts.reserve() parameter name (reserved_usd_micros)
   - Added independent sessions for worker and reaper threads
   - Fixed initial_balance timing (before reserve, not after)
   - Added session.commit() in thread finalize functions
   - Added fresh session for final balance verification

All 10 Reaper tests now passing:
- Zombie run detection and termination
- Minimum fee charging
- Race condition handling (Worker vs Reaper, Reaper vs Reaper)
- Exactly-once finalization guarantee

Technical details:
- Worker vs Reaper race test validates exactly-once semantics
- Budget calculations verified: charged = initial - final
- Session isolation prevents stale data in multi-threaded scenarios

+++ dpp/apps/api/dpp_api/db/models.py
@@ -3,7 +3,7 @@
 from datetime import date, datetime, timezone
 from typing import Optional
 
-from sqlalchemy import BIGINT, DATE, FLOAT, JSON, TEXT, TIMESTAMP, UUID, Index
+from sqlalchemy import BIGINT, DATE, FLOAT, JSON, TEXT, TIMESTAMP, UUID, Index, Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
 
@@ -187,7 +187,7 @@ class TenantPlan(Base):
 
     __tablename__ = "tenant_plans"
 
-    id: Mapped[int] = mapped_column(BIGINT, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     tenant_id: Mapped[str] = mapped_column(TEXT, nullable=False, index=True)
     plan_id: Mapped[str] = mapped_column(TEXT, nullable=False)
 
@@ -227,7 +227,7 @@ class TenantUsageDaily(Base):
 
     __tablename__ = "tenant_usage_daily"
 
-    id: Mapped[int] = mapped_column(BIGINT, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     tenant_id: Mapped[str] = mapped_column(TEXT, nullable=False, index=True)
     usage_date: Mapped[date] = mapped_column(DATE, nullable=False)
